refactor(api): :recycle: @trpc/server removed from api
@trpc/server related code moved from api to trpc package.
