+++ .prettierignore
@@ -1,0 +1,1 @@
+pnpm-lock.yaml
