+++ apps/expo/package.json
@@ -23,7 +23,6 @@
     "@tanstack/react-query": "^5.25.0",
     "@trpc/client": "11.0.0-rc.332",
     "@trpc/react-query": "11.0.0-rc.332",
-    "@trpc/server": "11.0.0-rc.332",
     "expo": "~50.0.7",
     "expo-constants": "~15.4.5",
     "expo-dev-client": "~3.3.8",
