+++ packages/api/package.json
@@ -24,7 +24,6 @@
     "@nestjs/core": "^10.0.0",
     "@nestjs/microservices": "^10.0.0",
     "@nestjs/websockets": "^10.0.0",
-    "@trpc/server": "11.0.0-rc.332",
     "zod": "^3.22.4"
   },
   "devDependencies": {
