+++ packages/api/src/index.ts
@@ -1,8 +1,11 @@
-import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
 import { NestFactory } from "@nestjs/core";
 import { AppModule } from "app.module";
 
-import type { Session } from "@acme/trpc";
+import type {
+  inferRouterInputs,
+  inferRouterOutputs,
+  Session,
+} from "@acme/trpc";
 import { createCallerFactory, createInnerContext } from "@acme/trpc";
 
 import type { AppRouter } from "./app.router";
@@ -40,22 +43,4 @@ export {
   AppRouterFactory,
   AppModule,
 };
-export type { Session, AppRouter, RouterInputs, RouterOutputs };
-
-/**
- * Below types needs to be exported.
- * Due to the issues coming in apps:
- * - "property collision with a built-in method"
- * - "router types not being inferred on apps"
- *
- * see: https://github.com/trpc/trpc/issues/5614
- */
-export {
-  type TRPC_ERROR_CODE_NUMBER,
-  type BuiltRouter,
-  type QueryProcedure,
-  type MutationProcedure,
-  type DecorateCreateRouterOptions,
-  type RouterCaller,
-  type Router,
-} from "@trpc/server/unstable-core-do-not-import";
+export type { AppRouter, RouterInputs, RouterOutputs, Session };
