+++ packages/trpc/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@nestjs/common": "^10.0.0",
     "@acme/logging": "workspace:^",
-    "@trpc/server": "11.0.0-next-beta.236",
+    "@trpc/server": "11.0.0-rc.332",
     "zod": "^3.22.4"
   },
   "devDependencies": {
