+++ packages/trpc/src/index.ts
@@ -12,3 +12,21 @@ export {
   inferRouterInputs,
   inferRouterOutputs,
 };
+
+/**
+ * Below types needs to be exported.
+ * Due to the issues coming in apps:
+ * - "property collision with a built-in method"
+ * - "router types not being inferred on apps"
+ *
+ * see: https://github.com/trpc/trpc/issues/5614
+ */
+export {
+  type TRPC_ERROR_CODE_NUMBER,
+  type BuiltRouter,
+  type QueryProcedure,
+  type MutationProcedure,
+  type DecorateCreateRouterOptions,
+  type RouterCaller,
+  type Router,
+} from "@trpc/server/unstable-core-do-not-import";
